Beta Reduction Step
Beta Reduction Step

In this exercise, you will implement beta reduction. This is a significant step in the project. Make sure you understand the material well enough first before implementing the step function.

As before, you have two options for getting started:

Use your solution to the previous exercise (Substitution or Complete) as the starting point, or
use the provided starter code, which is the model solution from the previous exercise with.
The starter code comes with the following files:

term.rs: contains the definition for Term and its methods.
lib.rs: makes the term.rs module part of the library crate.
main.rs: the REPL for testing and experimenting.
term/display.rs: pretty printer for Term.
term/util.rs: utilities for creating new terms.
term/parse.rs: parser for terms.
Your task is to implement the step and is_value methods in term.rs according to the material. Their type signatures are:

impl Term {
    pub fn step(self) -> Self {
        todo!()
    }

    pub fn is_value(&self) -> bool {
        todo!()
    }
}
See Definitions 6.4.1 and 6.4.2 for the exact specification. The grader only tests the step and is_value methods.

To start the REPL, run cargo run. Here is a sample from how the REPL works:

Enter :q or Ctrl+C to quit.
Entering an empty line steps the previous result.
> (fun x, fun y, x) (fun x, fun y, y) (fun x, x)
      ((𝜆 x. 𝜆 y. x) (𝜆 x. 𝜆 y. y)) (𝜆 x. x)
  --> (𝜆 y. 𝜆 x. 𝜆 y. y) (𝜆 x. x)
>
  --> 𝜆 x. 𝜆 y. y
> (fun x, fun y, x) (fun x, fun y, y) (fun x, fun y, x)
      ((𝜆 x. 𝜆 y. x) (𝜆 x. 𝜆 y. y)) (𝜆 x. 𝜆 y. x)
  --> (𝜆 y. 𝜆 x. 𝜆 y. y) (𝜆 x. 𝜆 y. x)
>
  --> 𝜆 x. 𝜆 y. y