Pairs
Pairs
25 / 25 points

Your task is to implement the dynamics and statics of pairs on top of STLC + ℤ + 𝟚 according to the evaluation and typing rules presented in the material. The starter code no longer contains the solution to the previous exercise, but only parsing and other utilities. From this point onwards you are expected to extend your version of the language with the new features.

Extend Term with the pairs and projections:
    /// A pair consisting of terms
    Pair(Box<Term>, Box<Term>),
    /// The first term in the pair
    Fst(Box<Term>),
    /// The second term in the pair
    Snd(Box<Term>),
Extend Type with products:
    /// Product type (Pair)
    Prod(Box<Type>, Box<Type>),
From the starter code, copy the parsing code (parse.rs) and pretty printer code (display.rs) from the respective directories for terms and types. The parsing code supports creating pairs with parenthesis, e.g. (1, 2) and the projection operators fst, snd. Note that applying fst, snd usually needs to be surrounded in parentheses. You may change this behavior in the parser if you wish.
The grader only tests the is_value, subst, step, multistep and type_check methods and does not test parsing or utilities.

Here is a sample from the REPL:

> let p = (fun x : Integer, (x, False), 2) in (fst p) 10
(10, False) :: ℤ × 𝟚
> let add = (fun p : (Integer, Integer), (fst p) + (snd p)) in add
𝜆 p : ℤ × ℤ. fst p + snd p :: ℤ × ℤ → ℤ
> let add = (fun p : (Integer, Integer), (fst p) + (snd p)) in add (2, 5)
7 :: ℤ