Declarations and Modules
Declarations and modules
30 / 30 points

Your task is to fill in the implementation of a programming language built on top of STLC + ℤ + 𝟚 + pairs + lists + sums + fix. Changes to terms and types are not required, as the exercise is about declarations, modules and imports.

Add pub mod module; to your lib.rs.
Copy module.rs and module from the starter.
Fill in the todo!()s in module.rs. See the function comments there for details.
The grader only tests the Declaration::to_term_fix and Module::to_term.

A new REPL is implemented in main.rs and is provided for debugging purposes. Running it with an argument, e.g. cargo run examples/std.stlc can be used import a path into the prelude of the REPL. Here is a sample from the REPL:

$ cargo run examples/std.stlc
examples/std.stlc λ reverse (cons 1 (cons 2 (nil Integer)))
[2, 1] :: [ℤ]

As an additional challenge, try writing an STLC++ module the following functions:

not :: 𝟚 → 𝟚
even :: ℤ → ℤ
len :: [ℤ] → ℤ
map :: (ℤ → ℤ) → [ℤ] → [ℤ]
zip :: [ℤ] → [ℤ] → [(ℤ → ℤ)]
unzip :: [(ℤ → ℤ)] → ([ℤ], [ℤ])
all :: (ℤ → 𝟚) → [ℤ] → 𝟚
sums :: [ℤ + ℤ] → (ℤ, ℤ)
However, no points are awarded for completing this challenge.